se elimina archivo URL y se pega carpetazo
