+++ src/domain/Login.ts
@@ -1,4 +1,17 @@
+export class LoginRequestDTO {
+  username: string;
+  password: string;
+    constructor (username: string, password: string ){
+        this.username = username;
+        this.password = password;
+    }
+  
+    static fromDto(us: string, pass: string): LoginRequestDTO {
+        return new LoginRequestDTO(us, pass);
+    }
+  }
 
+  
 export class LoginResponseDTO {
   public id: number = 0;
   public role: string = "";
@@ -24,4 +37,3 @@ export class LoginResponseDTO {
 }
 
 export const loginResponseDTO = new LoginResponseDTO(0, "");
-
