+++ src/services/moduleService.tsx
@@ -1,10 +1,10 @@
 import axios from 'axios' 
 import { Module } from "../domain/module";
-import { REST_SERVER_URL } from "./urls";
+import { URL_SERVIDOR_REST } from "../utils/config";
 
 class ModuleService {
     async getModules( id : number ) : Promise<Module[]>  {
-        const response = await axios.get(`${REST_SERVER_URL}/module/${id}/all`);
+        const response = await axios.get(`${URL_SERVIDOR_REST}/module/${id}/all`);
         return response.data
     }
 }
