--- src/services/urls.ts
@@ -1,1 +1,0 @@
-export const REST_SERVER_URL = 'http://localhost:8080'
